Change package name to lowercase
** Why are these changes being introduced:

Composer audit on this plugin flags the use of capital letters in the
package name, which causes problems downstream should anyone rely on
tools like composer diagnose.

** Relevant ticket(s):

Unticketed maintenance work

** How does this address that need:

This changes the package name to use only lowercase letters.

** Document any side effects to this change:

The version of the plugin bumps to 2.4.1-beta1 so I try a pre-release
on Github and consume it on our network.